Relocate MP types and methods from legacy_* files
Move all type definitions, const blocks, and var declarations from
8 legacy files into mptypes.go. Move all methods on those types into
mpmethods.go. This is Phase 1 of MP removal: types and their methods
must be relocated before legacy files can be deleted.

mptypes.go (795 lines): types, consts, vars only
mpmethods.go (614 lines): methods on those types

Source files:
- legacy_agent_structs.go (31 types/consts/vars — now empty)
- legacy_syncheddataengine.go (18 types + 11 methods — functions remain)
- legacy_hsyncengine.go (3 types — functions remain)
- legacy_combiner_chunk.go (5 types — functions remain)
- legacy_combiner_utils.go (3 vars — functions remain)
- legacy_hsync_utils.go (1 type — functions remain)
- legacy_db_combiner_edits.go (3 types — functions remain)
- legacy_apihandler_combiner_distrib.go (2 types — functions remain)

Build verified: all 6 binaries compile cleanly.
